DI Changes

+++ meetCodes/src/MeetCodes.Data/Models/MeetCode.cs
@@ -9,7 +9,8 @@ namespace MeetCodes.Data.Models
 {
     public class MeetCode:BaseEntity
     {
-        public virtual BigInteger Code { get; set; }
+
+        public virtual long Code { get; set; }
         public virtual string Description { get; set; }
         public virtual string Address1 { get; set; }
         public virtual string Address2 { get; set; }
